+++ packages/config/src/main.ts
@@ -10,11 +10,12 @@ import { resolveConfigPaths } from './files.js'
 import { getHeadersPath, addHeaders } from './headers.js'
 import { getInlineConfig } from './inline_config.js'
 import {
+  type ExtensionWithDev,
   EXTENSION_API_BASE_URL,
   EXTENSION_API_STAGING_BASE_URL,
-  mergeIntegrations,
-  NETLIFY_API_STAGING_BASE_URL,
-} from './integrations.js'
+  NETLIFY_API_STAGING_HOSTNAME,
+  normalizeAndMergeExtensions,
+} from './extensions.js'
 import { logResult } from './log/main.js'
 import { mergeConfigs } from './merge.js'
 import { normalizeBeforeConfigMerge, normalizeAfterConfigMerge } from './merge_normalize.js'
@@ -35,7 +36,7 @@ export type Config = {
   context: any
   env: any
   headersPath: any
-  integrations: any
+  integrations: ExtensionWithDev[]
   logs: any
   redirectsPath: any
   repositoryRoot: any
@@ -74,7 +75,7 @@ export const resolveConfig = async function (opts): Promise<Config> {
   }
 
   // TODO(kh): remove this mapping and get the extensionApiHost from the opts
-  const extensionApiBaseUrl = host?.includes(NETLIFY_API_STAGING_BASE_URL)
+  const extensionApiBaseUrl = host?.includes(NETLIFY_API_STAGING_HOSTNAME)
     ? EXTENSION_API_STAGING_BASE_URL
     : EXTENSION_API_BASE_URL
 
@@ -99,14 +100,14 @@ export const resolveConfig = async function (opts): Promise<Config> {
     featureFlags,
   } = await normalizeOpts(optsA)
 
-  let { siteInfo, accounts, integrations } = parsedCachedConfig || {}
+  let { siteInfo, accounts, integrations: extensions } = parsedCachedConfig || {}
 
   // If we have cached site info, we don't need to fetch it again
-  const useCachedSiteInfo = Boolean(featureFlags?.use_cached_site_info && siteInfo && accounts && integrations)
+  const useCachedSiteInfo = Boolean(featureFlags?.use_cached_site_info && siteInfo && accounts && extensions)
 
   // I'm adding some debug logging to see if the logic is working as expected
   if (featureFlags?.use_cached_site_info_logging) {
-    console.log('Checking site information', { useCachedSiteInfo, siteInfo, accounts, integrations })
+    console.log('Checking site information', { useCachedSiteInfo, siteInfo, accounts, extensions })
   }
 
   if (!useCachedSiteInfo) {
@@ -126,7 +127,7 @@ export const resolveConfig = async function (opts): Promise<Config> {
 
     siteInfo = updatedSiteInfo.siteInfo
     accounts = updatedSiteInfo.accounts
-    integrations = updatedSiteInfo.integrations
+    extensions = updatedSiteInfo.extensions
   }
 
   const { defaultConfig: defaultConfigA, baseRelDir: baseRelDirA } = parseDefaultConfig({
@@ -170,9 +171,9 @@ export const resolveConfig = async function (opts): Promise<Config> {
   // @todo Remove in the next major version.
   const configA = addLegacyFunctionsDirectory(config)
 
-  const updatedIntegrations = await handleAutoInstallExtensions({
+  const updatedExtensions = await handleAutoInstallExtensions({
     featureFlags,
-    integrations,
+    extensions,
     siteId,
     accountId,
     token,
@@ -184,15 +185,16 @@ export const resolveConfig = async function (opts): Promise<Config> {
     debug,
   })
 
-  const mergedIntegrations = await mergeIntegrations({
-    apiIntegrations: updatedIntegrations,
-    configIntegrations: configA.integrations,
-    context: context,
+  const mergedExtensions = normalizeAndMergeExtensions({
+    apiExtensions: updatedExtensions,
+    configExtensions: configA.integrations,
+    buildDir,
+    context,
   })
 
   const result = {
     siteInfo,
-    integrations: mergedIntegrations,
+    integrations: mergedExtensions,
     accounts,
     env,
     configPath,